Begin update procedure
The update procedure is quite long winded and this is just an initial check, following the packet structure
as declared in client.java -> anInt883 opcode 73 and 81. I need to organise the quirks the packets are
structured into - into methods which accomodate this. Perhaps map responses, or literal static classes
that respond to each packet. Not too sure, but something like this is definitely needed.