Reset byte structure
Did some reading on the client and I've made a little more progress,
we're upto the player list updating which is pretty monolithic and fairly confusing.
I'm going to continue to document it as to what is happening, hopefully
being able to write the correct server code for appearance updating and locale.