More research complete
So, it turns out the appearance updating actually waits for the update block flag 0x10 in method 107 to be run...
Which means the docs on rsps fandom are actually wrong. They also claim the final 2 bytes of the player list updating are
read as bytes, they're actually read as a short. Of course this is only a minor detail, but it's wrong nonetheless.
Onto block flag updating!