Player list updating query
Just performing a couple queries on the player list updating, it appears to start at 0? Or defaults some values for a player at index '0'?
Really not sure about this part, why would the client default a player at player index 0 with default values? It appears to read high to low,
i.e. 2046 -> 0. So maybe we should be starting at index 0...