Add initial load clause
Added an initial load clause, this needs turning into a game loop ideally, but due to the nature of node we're gonna
have to do this outside of the eventcycle which isn't really great, or come up with a clever way to force a wait as the client sends us data
per every game tick, and sends one per every packet received. I.e., tick speed / 2. We need it at around 500ms or w/e the ideal tick speed is
for Runescape.

Also added comment on the equipment loop for beards.