Attempt a game loop
Attempting to get the initial part of the game loop going, bit tired and can't focus so will complete this test tomorrow.
The premise is that after the login complete, it sets up an interval of 600ms emitting an event called "tick" to our gameloop event,
it also declares the call back right after this.

In the data event down, if our state is logged in, then we know we can freely just add to the buffer.
Not sure on the size of it, but ideally it should be how many times someone can click in 600ms lol + a couple for safety.

If my theory is right, then when I come to respond, I'll be able to write a giant packet with multiple frames responding to
each cached buffer opcode. Obviously I'll also need another buffer that actually stores the opcode + the payload to do anything.

In the memory tests, this uses next to nothing. I can imagine a few hundred players wouldn't do much either. And I'm clicking a fair few times
each run lol!

Time will tell, web dev 101 here:D