Fix architecture
Need to fix architecture in prep for handling and respond to client packets, it's one thing to read one and write one separately and another
to have it dynamically respond, as such, I'll need a pretty solid fluent extensible structure... I could approach this with having an abstract
'Packet' class which forces the packet to find its length and whatnot, that way when we instantiate a new packet we haven't gotta worry
about that code, it'll just have this.length ready for us? Questions further come when we consider creating packet block, with multiple frames.
So, got some thinking to do!