Refactor packets and login procedure
Decently large refactor, still need to do the PacketWriter, but that's more an implementation of a feature
rather than refactor, hence it's not in this push.

I've moved the login process into a statically handled class, if we log out, we're still gonna need the same instance
to handle our login... Or reconnection etc. I may even have it control it's own data event entirely dedicated to login
and then just keep checking the LoginHandler.LOGIN_STATE in the game data event as to whether we should
interpret game packets... Not sure.