Top left, bottom right todo
So, I'm getting lost in all the co-ords now lol. Will pick this up later, I need to draw a map out on my
whiteboard and properly calculate a good algorithm to send the correct positional packet as the correct time.
I still have a feeling them bytes have something to do with it...