Added 20,34 as our co-ords
This is so I can work out the pathing. SO I'm pretty sure the bytes are something to do with the pathing...
But, as far as I can tell there's not enough to get us all the way around.

In small objects like a 1x1 bush and clicking around it, it's giving us the correct amount. So maybe there's a limit?
And if that's the case, then each byte represents a tile, maybe...?

I'm thinking its either direction or a co-ordinate.
As the client write us this:
    outgoing.writeByte(waypointX[currentIndex] - x);
    outgoing.writeByte(waypointY[currentIndex] - y);
Which leads me to believe that we could simple take base/x/y + the byte +x/y and
pass that through the movement if statement block. At the end, remove the byte and repeat?

Really not sure though!