Design for pathing update
When packet 164 comes in and it's parsed, afterwards if we have bytes, parse them into literal co-ords.
Will maybe move this over toe p164 parse itself, but 248 and 98 will also use this. So I think checking if they're in a pathing state
after any of the 3 movement packets would be more suitable and DRY