Add IClient interface
This interface is very general, but if I understand NodeJS's TCP state, then we can store the socket (current accessed by that login)
into a player object to be pushed onto an array, and then when sending data, respond to that client? I'm not 100% and need to learn
a little more here.