Action ids and logout
I've added some of the basic action ids into a static class and the generic logout packet, now we need to create the client
objects for our Set() and respond to individual players currently connected within the Set() before we can 'log' a player out.

First step is to move the gameLoop to always be active when the server starts.