Taking a guess on updating others movements
I think I'll just try my guess and see if it works. The idea is that we store the playerid, move type and direction for EVERY player every 600ms.
For now I'm gonna update every player period and not check the region (in a proper situation I think we'd check the region and update only those in our region?).
Then in our P81, we adjust the update for other players movements to accodate this array but .filter(client => client.playerid != ourclient.playerid);, then summarise the length
and send the bits on a loop from this array to each client connected in a given region. (For now no regions but we get the idea).