playerCount in client others movement method
This field never updates other than in the place I need it to update before... lol. It's as though it's designed to skip over this? At least until our playerList has been updated with physical
player appearances... I'll come back to this with a freshmind tomorrow and try pinpoint
exactly what it's doing in the grand scheme of things.