Movement update progress
So, we can't update movements for others until they're on our playerlist. As they have no index to go ahead and update.
So initially we skip this as there's no movements for others to update even if they are logged in on the first tick because
they're not on our local players update list. We just add them to the list, but DON'T update their appearance. That way they're on the list
and their movement can be updated. Then appearance. It's a weird system if I'm interpretting it correctly, but it works.