End of Version1
So, after all I've done over these following 2 weeks, I've realised how the entire procedure of player updating fundamentally works now.
I think it'd be best to redesign the project in a 'version 2' and accomodate for all the findings I've found as well as document them thoroughly.
Next step will be to document it all as thoroughly as possible and write according to my own documentation. !