Remove gamecycle eventemitter
It simply just won't work for this kind of application, we have 0 way to remove a specific
listener, due to this, if a player logs out (even though it'll be doing nothing), that listener
will be listening forever. As for our loginHandler, that's fine. It's a single one per player
and will be wiped upon the socket being destroyed. I need to make my own system or
grab RxJS again. The standard eventemitter in Node is shit.