Add cycle subject
Decided to use RXJS for the cycle subject, may use an AsyncSubject though to truly force them to behave in sync.
Now I've got no idea of the memory consumption on a .subscribe() nor how to test it... So need to figure a way
to bench mark memory consumption for 2000 connections.