PacketReader & client -> player casting
Added a PacketReader class to read the lengths of variable packets and all the fixed length packets.
Added the casting to a player object within the login, makes more sense here. Also had to do this,
as the time to process the next data event was slightly off and this lead to the game packets being
out of sync. We now rely on the callback from our connectionEvent to change the data associated
with our current player. This will be used in saving, rather than a global connection variable.
Packet sync is now 100%.