Extension based bitwriting
I'm just trying this idea out, obviously the method names need changing but I think this looks clean.
It allows us to instantiate a bitwriter with an internal buffer, then it just converts and blast them all onto
a given buffer. Seems like a solid way of approaching this. We can write entire number and fix it to a specific length (which
is needed often in the bit writing of the RS protocols. I..e, value 1 to an 11 bit total.