Cleaned up the bitWriter
The BitWriter now has a beginning and end, it starts at an empty byte and ends on a full byte.
We can write numbers to a fixed length or their default length by passing 0 on writeNumber.
Added writeBit to make it explicitly clear what is happening. I think overall job well done,
feels clean and smart. Best of all it's simply based on an extension method system, so upon
finishing the used of it the compiler is going to clean up this block of memory, which further means
that if our Packet81 generator is static, we've not got to worry about stacks of BitWriters not being
garbage collected.