Multiple index writing support
Each local player can now write the indexes for each other client connected. Just need to modularise this, so we only send
an appearance update the once or on a change? Really not sure how to manage the state of interactions. I'll hardcode
appearance updates for now for each player temporarily.