Update player list with the local player
Updates the player list with the local player (and other players) player objects,
this allows us to emit data directly from one players action in their socket into an opposing players socket.
I.e., fluent interaction. Not that this is really need for getting movement going, but it for sure will be helpful
for combat.