Revert back, but add knowledge comment
So we've still had no luck in how or when we should append our masks or about the packet sizing in general. I feel strongly that the masks should NOT be included in the player list updating but don't know how to get around it. Will continue to read the code today and hopefully find a solution!