README update with issue
Well, we're stuck at a wall with the playerListUpdating aspect of the client, I've written a post and it's been a few days with no responses. Will await for guidance on how to approach this, I've traced the packetSize declaration and it's not dynamically determined throughout the packet. So I'm honestly at a wall.