Refactor
Small refactor in prep for handling incoming packets by bulk, as oppose to how we did it in v1, independently. I feel this will be a little more
straight forward, packets are read -> and then responded to. That's it. Not sure if any clauses in the future may stop me here, but we'll see I guess.