Packet architecture adjustment
I've decided the packet builders shouldn't actually flush the packets, this constrains us to sending individual packets...
If we want to block it together for set security of the data, we can't. So now I've added a 'getPacketNumber()' to them, it just
returns the buffer for the packet. This allows us to Buffer.concat(buffers) down the socket, which is much cleaner and is a single
stream emit rather than multiple.