Initial walk handling
Our packet array is shifted now rather than a for each, I personally think this looks cleaner.
We shoot the packets through a route handler method which inturn will call the correct handler
for a given packet. Currently 164 updates the players dx/dy and pathcoords. Need to think of a system
in which we progressively move, but also can variable bring masks in if needed.