P81 may need further adjustments
Left a fairly hefty comment on my ideal updates for P81 and the masks. I think the architecture may need changing to handle appending masks
independently as their 'own' buffer, and then update P81 with the new 'total' size. This means we'll treat the masks as their own packets in a way,
because they represent an operation and even have a bitmask opcode. So it makes sense, right?