ts-jest, jest, @types/jest, jest-junit, test function install
Our types fundamentally matter in the server, so using the babel7 transpile for our tests won't really be good enough. For example,
a Client -> a Player on login, we need to ensure the properties are set correctly etc. And when returning from a Player -> a Client,
we need to ensure this piece of memory is wiped before being reallocated. So ts-jest fits this perfectly.