Quickly test ts-jest functionality
Works perfectly fine, we can now test our handlers to ensure they behave how we want! We're kinda forced into test driven development here because when filtering
a list of players from numerous regions, we don't want to be constantly loggin in on different accounts or hardcoding a player to a region whilst we haven't got player
saving. So this is perfect.