Player list updating as and when needed, circ ref fix
So I've been having a few issues in this commit, circular references from the socket made things difficult. We could remove the socket itself... But that's a lot of faffing. I'd rather perform a direct check on the top level.
I found this lovely util in node's util's which parses bad refs into a string [TYPE *AMOUNT].