Player list updating
So for now, our player list updating portion is complete. An appearance render and that's it. We snapshot the previous players in range, and if the new call back determines it's changed,
we update it over again. Now we should be able to update their movement and it should delegate down to their 'otherY/X' within the player list updating if they enter our region/local area.